Move federation departure task out of main command
